Fix SEGV when XftFontOpenName() returns NULL.
The semantics of a callback set with XtSetTypeConverter() is to return
False only if there is no conversion at all.
In this case the converstion string is copied verbatim to the target.
Since the 'from' and 'to' types differ this doesn't make sense here at all.
Instead the target value needs to be set to NULL and the return needs
to be True.

Signed-off-by: Andreas Schwab <schwab@suse.de>
Reviewed-by: Egbert Eich <eich@freedesktop.org>